# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI E-Commerce Backend
  version: 0.1.0
paths:
...
  /api/v1/products:
    get:
      tags:
      - buyer
      summary: List Products
      operationId: list_products_api_v1_products_get
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: search_type
        in: query
        required: false
        schema:
          type: string
          default: keyword
          title: Search Type
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
      - name: shop_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Shop Id
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Price
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Price
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/products/{id}:
    get:
      tags:
      - buyer
      summary: Get Product
      operationId: get_product_api_v1_products__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/orders:
    post:
      tags:
      - buyer
      summary: Create Order
      operationId: create_order_api_v1_orders_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/orders/{id}:
    get:
      tags:
      - buyer
      summary: Get Order
      operationId: get_order_api_v1_orders__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/seller/shop:
    get:
      tags:
      - seller
      summary: Get Shop
      operationId: get_shop_api_v1_seller_shop_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
    put:
      tags:
      - seller
      summary: Update Shop
      operationId: update_shop_api_v1_seller_shop_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/seller/products:
    get:
      tags:
      - seller
      summary: List Seller Products
      operationId: list_seller_products_api_v1_seller_products_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - seller
      summary: Create Product
      operationId: create_product_api_v1_seller_products_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seller/products/{id}:
    put:
      tags:
      - seller
      summary: Update Product
      operationId: update_product_api_v1_seller_products__id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - seller
      summary: Delete Product
      operationId: delete_product_api_v1_seller_products__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seller/products/{id}/images:
    post:
      tags:
      - seller
      summary: Add Product Image
      operationId: add_product_image_api_v1_seller_products__id__images_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: url
        in: query
        required: true
        schema:
          type: string
          title: Url
      - name: is_primary
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Is Primary
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seller/products/images/{id}:
    delete:
      tags:
      - seller
      summary: Delete Product Image
      operationId: delete_product_image_api_v1_seller_products_images__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seller/orders:
    get:
      tags:
      - seller
      summary: List Orders
      operationId: list_orders_api_v1_seller_orders_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seller/orders/{id}:
    get:
      tags:
      - seller
      summary: Get Order
      operationId: get_order_api_v1_seller_orders__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - seller
      summary: Update Order Status
      operationId: update_order_status_api_v1_seller_orders__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/seller/chatbot-config:
    get:
      tags:
      - seller
      summary: Get Chatbot Config
      operationId: get_chatbot_config_api_v1_seller_chatbot_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
    patch:
      tags:
      - seller
      summary: Patch Chatbot Config
      operationId: patch_chatbot_config_api_v1_seller_chatbot_config_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatbotConfigIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/seller/stats:
    get:
      tags:
      - seller
      summary: Get Stats
      operationId: get_stats_api_v1_seller_stats_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatbotConfigIn:
      properties:
        api_key:
          type: string
          title: Api Key
        prompt:
          type: string
          title: Prompt
          default: ''
        template:
          type: string
          title: Template
          default: ''
        is_enabled:
          type: boolean
          title: Is Enabled
          default: false
      type: object
      required:
      - api_key
      title: ChatbotConfigIn
...
    CreateOrderIn:
      properties:
        shipping_address:
          type: string
          title: Shipping Address
      type: object
      required:
      - shipping_address
      title: CreateOrderIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderStatusIn:
      properties:
        status:
          type: string
          title: Status
      type: object
      required:
      - status
      title: OrderStatusIn
...
    ProductIn:
      properties:
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ''
        price:
          type: number
          title: Price
        stock:
          type: integer
          title: Stock
        category_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
      type: object
      required:
      - name
      - price
      - stock
      title: ProductIn
...
    ShopIn:
      properties:
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ''
      type: object
      required:
      - name
      title: ShopIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login